+++ src/load/azext_load/vendored_sdks/loadtesting/aio/_operations/_patch.py
@@ -68,6 +68,7 @@ class AsyncValidationCheckPoller(AsyncLoadTestingPollingMethod):
         self._polling_interval = interval
         self._status = None
         self._termination_statuses = [
+            "NOT_VALIDATED",
             "VALIDATION_SUCCESS",
             "VALIDATION_FAILURE",
             "VALIDATION_NOT_REQUIRED",